Airbnb_property_listing/requirements.txt
numpy==1.26.0
pandas==2.1.4
torch==2.1.2
torchvision==0.16
tensorboard==2.15.1
scikit-learn==1.3.2
PyYAML==5.4.1
joblib==1.0.1
seaborn==0.13.1
statsmodels==0.14.1
matplotlib==3.4.2
scipy==1.7.0